+++ Examples/BasicGame/include/BasicGame.hpp
@@ -2,6 +2,7 @@
 #include <Pyramid/Core/Game.hpp>
 #include <Pyramid/Graphics/Shader/Shader.hpp>
 #include <Pyramid/Graphics/Buffer/VertexArray.hpp>
+#include <Pyramid/Graphics/Texture.hpp> // Added for ITexture2D
 #include <memory> // For std::shared_ptr
 
 class BasicGame : public Pyramid::Game
@@ -18,4 +19,6 @@ protected:
 private:
     std::shared_ptr<Pyramid::IShader> m_shader;
     std::shared_ptr<Pyramid::IVertexArray> m_vertexArray;
+    std::shared_ptr<Pyramid::ITexture2D> m_texture; // Added
+    float m_time = 0.0f; 
 };